BSD 2-Clause "Simplified" License
This project is licensed under the BSD 2-Clause "Simplified" License

A permissive license that comes in two variants, the BSD 2-Clause and BSD 3-Clause.
Both have very minute differences to the MIT license.

Permissions

✓ Commercial use
✓ Modification
✓ Distribution
✓ Private use

Limitations

✗ Liability
✗ Warranty

Conditions

○ License and copyright notice

A copy of the license and copyright notice must be included with the software.